+++ _example/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"log"
+)
+
+func main() {
+	s := newServer("api/v1", 3000)
+	s.AddGlobalMiddleware(s.Logger)
+
+	log.Fatalln(s.ListenAndServe())
+}