+++ src/main.go
@@ -10,18 +10,20 @@ import (
 func main() {
 	// Intro.
 	fmt.Println()
-	fmt.Println("Near Gothic")
+	fmt.Println("NEAR GOTHIC")
 	fmt.Println("Generate Golang repository code directly from Postgres")
 	fmt.Println()
 
 	// Command arguments.
 	var a = argsParser.New(os.Args)
-	a.Example = "-w -folder ~/example/repo -module kcartlidge/app/data -env DB_CONNSTR -schema example"
-	a.AddFlag("w", false, false, "overwrite any existing destination folder")
+	a.Example = "-w -env DB_CONNSTR -schema example -module kcartlidge/api -folder ~/example -repo repo"
+	a.AddFlag("w", false, false, "overwrite any existing destination folder?")
+
 	a.AddValue("env", false, "DB_CONNSTR", "connection string environment variable")
 	a.AddValue("schema", false, "public", "the Postgres database schema to scan")
-	a.AddValue("folder", true, "", "destination folder, either relative or absolute")
 	a.AddValue("module", true, "", "the Go module for created code (eg `kcartlidge/app/data`)")
+	a.AddValue("folder", true, "", "destination folder, either relative or absolute")
+	a.AddValue("repo", true, "", "repository subfolder name")
 	a.ShowUsage()
 	a.Parse()
 	if a.HasIssues {
@@ -33,14 +35,18 @@ func main() {
 	overwrite := a.Flags["w"]
 	env := a.Values["env"]
 	schema := a.Values["schema"]
-	folder := a.Values["folder"]
 	module := a.Values["module"]
+	folder := a.Values["folder"]
+	repoName := a.Values["repo"]
+	fmt.Println()
+	fmt.Println("Overwrite existing?  :", overwrite)
 	fmt.Println()
 	fmt.Println("Environment variable :", env)
 	fmt.Println("Database schema      :", schema)
-	fmt.Println("Destination folder   :", folder)
 	fmt.Println("Go module name       :", module)
-	fmt.Println("Overwrite existing?  :", overwrite)
+	fmt.Println("Destination folder   :", folder)
+	fmt.Println("Repo package name    :", repoName)
+	fmt.Println()
 	fmt.Println()
 
 	// Fetch the connection string from the env, and test it.
@@ -57,17 +63,18 @@ func main() {
 
 	// Create the output.
 	fmt.Println()
-	w := NewWriter(folder, module, a.CommandLine, env, s.Schema)
+	w := NewWriter(folder, module, a.CommandLine, env, s.Schema, repoName)
 	exists, err := Exists(w.topFolder)
 	check(err)
-	if !exists || overwrite {
-		w.WriteStuff()
-	} else {
-		check(errors.New("Stuff exists at the output location (-w overwrites)"))
+	if exists && !overwrite {
+		check(errors.New("The output folder exists (-w overwrites)"))
 	}
+	w.WriteStuff()
 
 	// Done.
+	fmt.Println()
 	fmt.Println("Done")
+	fmt.Println()
 }
 
 // If there is an error, display it and quit.